Objects emit noise on pickup
Plays a sound when picking up an object.........its the same one per object but it's something. If you need to change the noise the object makes, change it in the "Eddie_First_Person_Controller" blueprint since its the object that picks up objects and plays the sound when objects get picked up.

P.S. Nevermind on the Pierogies, they were fuckin delish with this indian cucumber sauce....just...damn....